Move cards to shared folder
