+++ front/app/components/admin/GraphCards/DeviceTypesCard/index.tsx
