+++ front/app/components/admin/GraphCards/DeviceTypesCard/messages.ts
