+++ front/app/components/admin/GraphCards/DeviceTypesCard/renderTooltip.tsx
