+++ front/app/components/admin/GraphCards/DeviceTypesCard/translations.ts
