+++ front/app/components/admin/GraphCards/DeviceTypesCard/useDeviceTypes.ts
